+++ src/cleaner.rs
@@ -27,8 +27,6 @@ use std::io;
 use svgdom;
 use svgdom::{Document, ParseOptions, WriteOptions, WriteBuffer, ElementId};
 
-use svgdom_utils;
-
 use options::Options;
 use task::*;
 use error;
@@ -58,7 +56,7 @@ pub fn clean_doc(doc: &Document, options: &Options, opt: &WriteOptions)
     // Prepare our document.
     // This methods is not optional.
     utils::resolve_gradient_attributes(doc)?;
-    svgdom_utils::resolve_inherit(doc)?;
+    svgdom::postproc::resolve_inherit(doc)?;
     fix_invalid_attributes(doc);
     group_defs(doc);