+++ src/lib.rs
@@ -24,7 +24,6 @@
 #[macro_use] extern crate clap;
 
 extern crate svgdom;
-extern crate svgdom_utils;
 
 pub use svgdom::{ParseOptions, WriteOptions};