+++ src/task/apply_transforms/gradients.rs
@@ -23,16 +23,16 @@
 use task::short::{EId, AId};
 use super::utils;
 
-use svgdom::Document;
+use svgdom::{Document, ElementType};
 
 pub fn apply_transform_to_gradients(doc: &Document) {
     let iter = doc.descendants().svg()
-                  .filter(|n| ::task::is_gradient(n))
+                  .filter(|n| n.is_gradient())
                   .filter(|n| n.has_attribute(AId::GradientTransform));
 
     for node in iter {
         {
-            let flag = node.linked_nodes().any(|n| ::task::is_gradient(&n));
+            let flag = node.linked_nodes().any(|n| n.is_gradient());
 
             if flag {
                 // We can apply transform only to gradients that doesn't used by